Adds PDF generation functionality for DDT documents
Implements the ability to generate PDF documents for DDTs (Documenti di Trasporto).

This includes adding the `pdfmake` library and its types as dependencies,
and integrating it into the DDT management page to create and download
PDF representations of DDT data.
Also fixes a bug with DDT id in update status function