Fetches and displays customer references
Adds functionality to display customer references in the customer details panel.

Fetches existing customer data from the API to populate the "Referred By" autocomplete field when adding new customers.

Displays a list of customers referred by the selected customer and a display of the customer who referred the selected customer in a new tab in the customer details panel.